OvrIntegration: Hide HmdStatusFlag and HmdStatusFlags from API.
